Bug 453347: NPE when hovering over run dialog
The location passed to ScriptService.getScriptType can be null if
the IFile being resolved is not a physical file on the user's machine.

As this code is called in the property tester it means that you
can get these NPEs randomly when doing things like hovering over
Run/Debug buttons, browsing through context menu, or anything else
that causes property tester to run.

In addition, I have added appropriate null checks to the return
values of all calls to IScriptService.getScriptType(String) and
ResourceTools.toAbsoluteLocation(Object, Object).

Change-Id: I5769e9832ea9b0e0f6a7246876164309f8c297d2
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>