Bug 499667 - Add support for completions of Python in Script Console
Use __main__ dictionary as globals for executed code.

The __main__ dictionary is the scope in which top-level code executes in
Python [1]. It makes sense that the actual toplevel scope for PY4J
scripts should be the file being executed rather than the EASE Py4j
framework.

This commit is necessary to have variables from the top level scope
available in code completion library.

[1] https://docs.python.org/3/library/__main__.html

Change-Id: I0f220c16c90fc0f8824a728be2ae527f5847c38e
Signed-off-by: Martin Kloesch <martin.kloesch@gmail.com>