Bug 528197: Option to ignore PYTHON* envs (-E command line)
Allow user to set an independent python to launch with independent
environment variables. When launching default python off the PATH
it is logical to use such variables, but when using a fully
constrained environment where the python executable is provided,
all the PYTHON* should be ignored to ensure that they do not
interfere.

Change-Id: I233f218a995af6f5ceec828af1eb023d10f864b4
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>