Bug 530037: Progressbar in Unit Test view is not read for error/failures
According to the ProgressBar javadoc, the state is the preferred way to
change the color of the progress bar. Now the setState method of the
progress bar is called in order to change its state and so its color.
The ERROR state is used for ERRORS and FAILURES (same behavior as in
JUnit).

Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=530037
Change-Id: Ic127d50b0eb4f238d15bdb13de02f558c1b7c380
Signed-off-by: Martin SCHREIBER <m.schreiber@bachmann.info>